+++ HabboHotel/Rooms/Room.cs
@@ -120,6 +120,8 @@ public class Room : RoomData
     private readonly IRoomRollerService _roomRollerService;
     private readonly IRoomItemInventoryService _roomItemInventoryService;
     private readonly IRoomItemUpdateQueueService _roomItemUpdateQueueService;
+    private readonly IRoomItemLoadService _roomItemLoadService;
+    private readonly IRoomItemRemovalService _roomItemRemovalService;
     private readonly IRoomService _roomService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
@@ -148,7 +150,7 @@ public class Room : RoomData
     private readonly IRoomManager _roomManager;
     private readonly ILanguageManager _languageManager;
 
-    public Room(RoomData data, IGameClientManager clientManager, IDatabase database, IItemLoader itemLoader, IRoomItemPersistenceService roomItemPersistenceService, IRoomItemPlacementValidatorService roomItemPlacementValidatorService, IRoomItemPlacementPersistenceService roomItemPlacementPersistenceService, IRoomRollerService roomRollerService, IRoomItemInventoryService roomItemInventoryService, IRoomItemUpdateQueueService roomItemUpdateQueueService, IGroupManager groupManager, IRoomService roomService, IChatManager chatManager, IBotManager botManager, IAchievementService achievementService, IQuestService questService, ICacheManager cacheManager, ILanguageManager languageManager, IItemTeleporterFinder itemTeleporterFinder, IItemHopperFinder itemHopperFinder, IBadgeManager badgeManager, IUserDataFactory userDataFactory, IRoomManager roomManager, ILoggerFactory loggerFactory)
+    public Room(RoomData data, IGameClientManager clientManager, IDatabase database, IItemLoader itemLoader, IRoomItemPersistenceService roomItemPersistenceService, IRoomItemPlacementValidatorService roomItemPlacementValidatorService, IRoomItemPlacementPersistenceService roomItemPlacementPersistenceService, IRoomRollerService roomRollerService, IRoomItemInventoryService roomItemInventoryService, IRoomItemUpdateQueueService roomItemUpdateQueueService, IRoomItemLoadService roomItemLoadService, IRoomItemRemovalService roomItemRemovalService, IGroupManager groupManager, IRoomService roomService, IChatManager chatManager, IBotManager botManager, IAchievementService achievementService, IQuestService questService, ICacheManager cacheManager, ILanguageManager languageManager, IItemTeleporterFinder itemTeleporterFinder, IItemHopperFinder itemHopperFinder, IBadgeManager badgeManager, IUserDataFactory userDataFactory, IRoomManager roomManager, ILoggerFactory loggerFactory)
         : base(data)
     {
         _clientManager = clientManager;
@@ -160,6 +162,8 @@ public class Room : RoomData
         _roomRollerService = roomRollerService;
         _roomItemInventoryService = roomItemInventoryService;
         _roomItemUpdateQueueService = roomItemUpdateQueueService;
+        _roomItemLoadService = roomItemLoadService;
+        _roomItemRemovalService = roomItemRemovalService;
         _groupManager = groupManager;
         _roomService = roomService;
         _chatManager = chatManager;
@@ -181,7 +185,7 @@ public class Room : RoomData
         MutedUsers = new();
         _tents = new();
         _gamemap = new(this, data.Model);
-        _roomItemHandling = new(this, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService);
+        _roomItemHandling = new(this, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService);
         _roomUserManager = new(this, clientManager, database, groupManager);
         _filterComponent = new(this);
         _wiredComponent = new(this, loggerFactory.CreateLogger<WiredComponent>());
@@ -244,7 +248,7 @@ public class Room : RoomData
 
     public RoomItemHandling GetRoomItemHandler()
     {
-        if (_roomItemHandling == null) _roomItemHandling = new(this, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService);
+        if (_roomItemHandling == null) _roomItemHandling = new(this, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService);
         return _roomItemHandling;
     }