+++ HabboHotel/Rooms/RoomManager.cs
@@ -34,6 +34,8 @@ public class RoomManager : IRoomManager
     private readonly IRoomRollerService _roomRollerService;
     private readonly IRoomItemInventoryService _roomItemInventoryService;
     private readonly IRoomItemUpdateQueueService _roomItemUpdateQueueService;
+    private readonly IRoomItemLoadService _roomItemLoadService;
+    private readonly IRoomItemRemovalService _roomItemRemovalService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
     private readonly IRoomService _roomService;
@@ -61,6 +63,8 @@ public class RoomManager : IRoomManager
         IRoomRollerService roomRollerService,
         IRoomItemInventoryService roomItemInventoryService,
         IRoomItemUpdateQueueService roomItemUpdateQueueService,
+        IRoomItemLoadService roomItemLoadService,
+        IRoomItemRemovalService roomItemRemovalService,
         IGameClientManager gameClientManager,
         IGroupManager groupManager,
         IRoomService roomService,
@@ -86,6 +90,8 @@ public class RoomManager : IRoomManager
         _roomRollerService = roomRollerService;
         _roomItemInventoryService = roomItemInventoryService;
         _roomItemUpdateQueueService = roomItemUpdateQueueService;
+        _roomItemLoadService = roomItemLoadService;
+        _roomItemRemovalService = roomItemRemovalService;
         _clientManager = gameClientManager;
         _groupManager = groupManager;
         _roomService = roomService;
@@ -227,7 +233,7 @@ public class RoomManager : IRoomManager
 
     private Room CreateRoomInstance(RoomData data)
     {
-        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
+        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
     }
 
     private static void DisposeRoom(Room room) => room.Dispose();